refactor: moved MainResult
