+++ src/components/Main/Boards/MainResults/MainResults.tsx
@@ -1,0 +1,27 @@
+import Box from '@mui/joy/Box';
+import CircularProgress from '@mui/joy/CircularProgress';
+
+import { useAppSelector } from '../../../../store/hooks';
+import { useGetBoardsByUserIdQuery } from '../../../../store/slices/boards/boardsApi';
+
+import { DialogEditBoard } from '../../../DialogEditBoard/DialogEditBoard';
+import { BoardCard } from '../BoardCard';
+
+export const MainResults = () => {
+  const { isOpenedDialogEditBoard } = useAppSelector((state) => state.boards);
+  const { id: userId } = useAppSelector((state) => state.user);
+
+  const { data: boards, isLoading } = useGetBoardsByUserIdQuery(userId);
+
+  const cards = boards?.map((board) => <BoardCard key={board._id} board={board} />);
+
+  return (
+    <>
+      <Box sx={{ display: 'flex', flexDirection: 'row', flexWrap: 'wrap', justifyContent: 'center', gap: 2 }}>
+        {isLoading ? <CircularProgress color="primary" size="lg" value={25} variant="soft" /> : cards}
+      </Box>
+
+      {isOpenedDialogEditBoard ? <DialogEditBoard /> : null}
+    </>
+  );
+};
